api/requirements.txt
Django==2.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
